% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklistRegions-data.R
\docType{data}
\name{blacklistRegions}
\alias{blacklistRegions}
\title{Data frame of blacklisted SNP regions}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(blacklistRegions)
}
\description{
Data summarizing blacklisted SNP regions based on ENCODE
}
\keyword{datasets}
