% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseSegmentsWithSegs-function.R
\name{phaseSegmentsWithSegs}
\alias{phaseSegmentsWithSegs}
\title{Phasing SNPs on segments across samples}
\usage{
phaseSegmentsWithSegs(snpTable, segTable, samples = NULL)
}
\arguments{
\item{snpTable}{A table of SNPs with their positions.}

\item{segTable}{A table of segments to consider including chromosome start and stop positions as well as a BAF value per segment.}

\item{samples}{A vector of sample names to include for phasing.
Defaults to NULL in which case all samples are considered.}
}
\value{
A table similar to the SNPtable provided as input, but also including
the segment the SNPs can be found on as well as classification whether they
are likely part of allele A or B.
}
\description{
This function phases SNPs located on the same segment across samples
}
