% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotPhasedSNPs}
\alias{plotPhasedSNPs}
\title{Plotting phasing of SNPs across the genome for multiple samples}
\usage{
plotPhasedSNPs(
  snpTable,
  genome,
  samples = NULL,
  col.palette = setNames(c("#f1a340", "#998ec3", "#bdbdbd"), c("top", "bottom", "mixed"))
)
}
\arguments{
\item{snpTable}{A table of SNPs with their positions along with their characterization.}

\item{genome}{A genome version to use, for example BSgenome.Hsapiens.UCSC.hg19.}

\item{samples}{A vector of sample names to include for plotting in the order they appear in.
Defaults to NULL in which case all samples are considered.}

\item{col.palette}{A named vector of colors for plotting the SNPs.
Defaults to orange and purple for allele A and B, respectively.}
}
\value{
A ggplot object that can be plotted directly or saved as a pdf.
}
\description{
This function plots the phased SNPs for all samples of interest.
}
