% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problematicLoci-data.R
\docType{data}
\name{problematicLoci}
\alias{problematicLoci}
\title{Data frame of problematic SNP loci}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(problematicLoci)
}
\description{
Data summarizing loci of problematic SNPs
}
\keyword{datasets}
