% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingSCNA-function.R
\name{testingSCNA}
\alias{testingSCNA}
\title{Testing presence of allelic imbalance using BAF estimates of SNPs}
\usage{
testingSCNA(
  phasingTable,
  newSegTable,
  seg.totalSNPs = 8,
  seg.nSNPs = 4,
  testMSAI = TRUE
)
}
\arguments{
\item{phasingTable}{A table of SNPs with their positions along with their characterization.
Additionally, their BAF and the segment they are assigned to are also included.}

\item{newSegTable}{A table of segments for all regions with chromosome, start, end, width.
Additionally, the segment ID as well as the BAF of the segment
and major and minor copy number are included.}

\item{seg.totalSNPs}{An integer used to filter minimum size of segments to test.
Defaults to 8 meaning segments with less than 8 SNPs are not tested.}

\item{seg.nSNPs}{An integer used to filter minimum size of SNPs per allele.
Defaults to 4 meaning at least 4 SNPs need to be characterized per allele to test segment.}

\item{testMSAI}{A flag whether called or rescued segments should be tested for MSAI.
Defaults to TRUE meaning MSAI is tested in all called or rescued segments.}
}
\value{
A table similar to the phasingTable provided as input, but also including
the characterization of whether allelic imbalance is detected/rescued in the segment.
}
\description{
This function tests whether for a given segment and region a BAF seperation of SNPs can be detected
based on previous characterization of SNPs as either allele A or B.
}
